+++ packages/langchain/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/langchain
 
+## 2.0.0-beta.87
+
+### Patch Changes
+
+- Updated dependencies [ca13d26]
+  - ai@6.0.0-beta.87
+
+## 2.0.0-beta.86
+
+### Patch Changes
+
+- Updated dependencies [36b175c]
+  - ai@6.0.0-beta.86
+
 ## 2.0.0-beta.85
 
 ### Patch Changes